==> notebook1.ipynb <==
#Script que usa a tool Pydriller para extrair arquivos envolvidos nas refatorações

import pandas as pd
import os
import json
from pydriller import Repository


dataset="eclipse"
with open(f'.\\..\\2_REFATORAÇÕES\\{dataset}_refactorings_final.json', 'r') as f:
    myJSON = json.load(f)

def isClassLevel(element):
    if str(element).find('class') != -1 or str(element).find('Class') != -1:
      return True
    else: 
        return False

def extractFiles():
    for r in range(0, 200):
        count=0
        for c in range(len(myJSON[r])):
            print(len(myJSON[r]))
            os.system('cls')
            print(f'repo:{r}/{len(myJSON)}')
            print(f'commit:{c}/{len(myJSON[r])}')
            if isClassLevel(myJSON[r][c]['element']):
                if count>300:
                    break
                count=count+1
                print(f"c {c}, {myJSON[r][c]['element']}")
                print(count)
                try:
                    url=f'.\\..\\1DATASET\\projetos\\{dataset}\\'+myJSON[r][c]['folder'] 
                    singleCommit = myJSON[r][c]['commit']
        
                    for commit in Repository(url, single=singleCommit).traverse_commits():
                        os.system(f'md .\\abfiles\\{dataset}\\r{r}\\c{c}\\before')
                        os.system(f'md .\\abfiles\\{dataset}\\r{r}\\c{c}\\after')
                        for f in commit.modified_files:
                            try: 
                                with open(f'.\\abfiles\\{dataset}\\r{r}\\c{c}\\before\\{f.filename}', 'w') as arquivo1:
                                    arquivo1.write(f.source_code_before) 
                            except: pass
                            try: 
                                with open(f'.\\abfiles\\{dataset}\\r{r}\\c{c}\\after\\{f.filename}', 'w') as arquivo2:
                                    arquivo2.write(f.source_code)
                            except: pass
                except: pass        

extractFiles()

#Obter arquivos envolvidos nas refatorações

from pydriller import Repository
import os
import json

#Parte 1 
dataset="apache"
x = open(f'.\\..\\2_REFATORAÇÕES\\{dataset}_refactorings_final.json')
myJSON = json.load(x)

#Parte 2 - Iterar sobre os projetos
for projeto in myJSON:
    # Acessar as refatorações
    for refatoracao in projeto["refactorings"]:
        # Verificar se a refatoração é um dicionário e contém a chave 'commits'
        if isinstance(refatoracao, dict) and "commits" in refatoracao:
            # Acessar os commits com refatorações
            for commitRow in refatoracao["commits"]:
                commit = commitRow["sha1"]
                
                #Acessar commits com refatorações apenas Class/Method
                for ref in commitRow['refactorings']:
                    if ref['type'].lower().find('class') != -1 or ref['type'].lower().find('method') != -1:

                        # Criar uma pasta para cada commit
                        folderDestination=f'{dataset}/{commit}'
                        os.makedirs(folderDestination, exist_ok=True)
                        
                        # Criar pastas 'before' e 'after' dentro da pasta do commit
                        os.makedirs(os.path.join(folderDestination, 'before'), exist_ok=True)
                        os.makedirs(os.path.join(folderDestination, 'after'), exist_ok=True)
                        
                        url=f'./../1_PROJETOS/projects/{dataset}/'+projeto["project"]
                        # Obter os arquivos modificados em cada commit usando PyDriller
                        
                        for c in Repository(url, single=commit).traverse_commits():
                            display(f'{projeto["project"]}: {commit}')
                            for modification in c.modified_files:
                                #verificar se o arquivo está dentro da lista de arquivos envolvidos na refatoração  
                                file = modification.filename
                                if ref['description'].lower().find(file.split('.')[0].lower()) != -1:

                                    # Verifica se o arquivo é java
                                    extensao = file.split('.')[-1]
                                    if (extensao == 'java'):
                                        print(f'{projeto["project"]}: {commit} - {file}')
                                        try:
                                            with open(os.path.join(folderDestination, 'before', file), 'w') as file1:
                                                file1.write(modification.source_code_before)
                                        except:
                                            pass
                                        try:
                                            with open(os.path.join(folderDestination, 'after', file), 'w') as file2:
                                                file2.write(modification.source_code)
                                        except:
                                            pass
                                        

#Obter arquivos envolvidos nas refatorações

from pydriller import Repository
import os
import json

#Parte 1 
dataset="random"
x = open(f'.\\..\\2_REFATORAÇÕES\\{dataset}_refactorings_final.json')
myJSON = json.load(x)

#Parte 2 - Iterar sobre os projetos
for projeto in myJSON:
    # Acessar as refatorações
    for refatoracao in projeto["refactorings"]:
        # Verificar se a refatoração é um dicionário e contém a chave 'commits'
        if isinstance(refatoracao, dict) and "commits" in refatoracao:
            # Acessar os commits com refatorações
            for commitRow in refatoracao["commits"]:
                commit = commitRow["sha1"]
                
                #Acessar commits com refatorações apenas Class/Method
                for ref in commitRow['refactorings']:
                    if ref['type'].lower().find('class') != -1 or ref['type'].lower().find('method') != -1:

                        # Criar uma pasta para cada commit
                        folderDestination=f'{dataset}/{commit}'
                        os.makedirs(folderDestination, exist_ok=True)
                        
                        # Criar pastas 'before' e 'after' dentro da pasta do commit
                        os.makedirs(os.path.join(folderDestination, 'before'), exist_ok=True)
                        os.makedirs(os.path.join(folderDestination, 'after'), exist_ok=True)
                        
                        url=f'./../1_PROJETOS/projects/{dataset}/'+projeto["project"]
                        # Obter os arquivos modificados em cada commit usando PyDriller
                        
                        #itera as mudanças dentro de uma refatoração
                        for c in Repository(url, single=commit).traverse_commits():
                            
                            #itera os arquivos envolvidos dentro de uma mudança
                            for modification in c.modified_files:
                                #verificar se o arquivo está dentro da lista de arquivos envolvidos na refatoração  
                                file = modification.filename
                                if ref['description'].lower().find(file.split('.')[0].lower()) != -1:

                                    # Verifica se o arquivo é java
                                    extensao = file.split('.')[-1]
                                    if (extensao == 'java'):
                                        clear_output()
                                        print(f'{projeto["project"] }: {commit}')
                                        try:
                                            with open(os.path.join(folderDestination, 'before', file), 'w') as file1:
                                                file1.write(modification.source_code_before)
                                        except:
                                            pass
                                        try:
                                            with open(os.path.join(folderDestination, 'after', file), 'w') as file2:
                                                file2.write(modification.source_code)
                                        except:
                                            pass
                                        

#Contando refatoações de classes e métodos
 
import json
from collections import Counter

dataset="random"
x = open(f'.\\..\\2_REFATORAÇÕES\\{dataset}_refactorings_final.json')
myJSON = json.load(x)


array = []
# Iterar sobre os projetos
for projeto in myJSON:
    # Acessar as refatorações
    for refatoracao in projeto["refactorings"]:
        # Verificar se a refatoração é um dicionário e contém a chave 'commits'
        if isinstance(refatoracao, dict) and "commits" in refatoracao:
            # Acessar os commits
            for row in refatoracao["commits"]:
                for r in row['refactorings']:
                    array.append(r['type'])
                # print(commit)  # Isso imprimirá cada commit
                # Iterar sobre os projetos
#imprimir array com suas respectivas quantidades
type_count = Counter(array)

sum=0
for type, count in type_count.items():
    if type.lower().find('class') != -1 or type.lower().find('method') != -1:
        print(f"Type: {type}, Count: {count}")
        sum=sum+count


print(sum)

from IPython.display import clear_output
import time

for i in range(5):
    clear_output()
    print(i)
