==> notebook3_PMD.ipynb <==
import subprocess
import os
import pandas as pd

dataset='apache'
type_out='issues'
ruleset='C:/PMD/bin/rules.xml'
src_folder= f'./../3_FILES/{dataset}/'
metrics=[]

def run_cmd(cmd):
    process = subprocess.Popen(cmd, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
    stdout, stderr = process.communicate()
    return stdout, stderr

def run_PMD(target_folder):
    output=f'output_{type_out}_PMD.csv'
    stdout, stderr = run_cmd(f'pmd check -R {ruleset} -d {target_folder} -f csv -r {target_folder}_{output}')  # substitua "dir" pelo comando que deseja executar
    
    
    if stdout:
        # print("Output:", stdout.decode('utf-8', errors='ignore'))
        metrics_output = pd.read_csv(f'{target_folder}_{output}')
        return len(metrics_output)
    if stderr:
        print("Error:", stderr.decode('utf-8', errors='ignore'))

#loop para percorrer todos os nomes de pastas de uma pasta
for i, commit in enumerate(os.listdir(src_folder)):
    cs_before = run_PMD(f'{src_folder}{commit}/before') 
    cs_after = run_PMD(f'{src_folder}{commit}/after') 
    metrics.append({'commit':commit, 'cs_before':cs_before, 'cs_after':cs_after})
    print(f'iteration {i} in total {len(os.listdir(src_folder))} ')

#salvar metrics em json como nome do dataset
df = pd.DataFrame(metrics)
df.to_csv(f'./4_output_pmd/{dataset}_{type_out}_pmd.csv', index=False)





import pandas as pd
dataset='apache'
df = pd.DataFrame([{'commit': '1', 'cs_before': 1, 'cs_after': 2}, {'commit': '2', 'cs_before': 3, 'cs_after': 4}])
df.to_csv(f'./4_output_pmd/{dataset}_pmd.csv', index=False)
