==> notebook7_mergeSS.ipynb <==
### Notebook que faz merge entre features SS, comparando nome do arquivo e commit.

import pandas as pd
import numpy as np
import os


dataset='apache'
# 1. Ler o arquivo CSV contendo o dataset
df_input1 = pd.read_csv(f'./5_output_merge2/{dataset}/datasetClass_with_foc_pmd_email.csv')

# 2. Abrir o outro arquivo CSV com as colunas 'commit', e 'ss'
df_ss = pd.read_csv(f'./7_output_scoreSimiliarity/{dataset}_score.csv')


df_input1['classMetric_file2']= df_input1['classMetric_file'].apply(lambda x: x.split('\\')[-1])
df_ss['file2']=df_ss['file'].apply(lambda x: x.split('\\')[-1])

# Fazendo o merge
df_merged = pd.merge(df_input1, df_ss, left_on=['classMetric_file2', 'commit'], right_on=['file2', 'commit'], how='left')

directory = f'./8_output_merge3/{dataset}'
if not os.path.exists(directory):
    os.makedirs(directory)
    
df_merged.to_csv(f'{directory}/datasetClass_with_foc_pmd_email_ss.csv', index=False)
